+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Removed pagination from all tables; it took up extra room and was a nuisance
 - Added keybindings for moving forward/back across the tab bar; changed default for closing the tab
 - Added module setting to show the type of each node in the Setting tree in hierarchy mode.
+- Consolidated "strong start" onto the Session notes tab
 - Minor UI cleanup 
 - Fixed bug where bookmarks weren't deleted when the underlying content was
 - Fixed bug where campaigns created by right-clicking in the Setting directory didn't immediately show up in the Campaign directory